perf: significantly speed up constraint matching performance (#171)
Removes constraint match total map and indictment map from the scope of
incremental calculation.
They are only built when requested by the user.

This leads to significant decreases in hash map usage during incremental
processing,
as well as considerably reduced allocation rates.
Enabling constraint matches is now less than 10 % slower than not, on
average. Used to be around 50 %.

This was achieved while at the same time improving the time it takes for
`SolutionManager` to explain a score.
This is due to a plethora of other minor optimizations throughout the
incremental operations.