fix: Make KOptListMoveSelector work when moveThreadCount != NONE
The issue came from the fact MultipleDelegateList returns a new
instance of the rebased list on every call. This is problematic,
since each FlipSublistAction expects rebase to return the same
list across mulitple calls. When different instances are used,
the FlipSublistAction modify different lists, causing the Undo
move to be incorrect. (Since the list the FlipSublistAction operation
on only has its flip incorporated in it instead of all flips).
Since a new MultipleDelegateList is created for each move, we can
safely store a cached result in one of its fields.