fix: avoid metaspace leak by no longer using the LambdaMetaFactory (#257)
The leak came from the fact that once the factory creates the lambda, it
becomes pinned to a classloader and there is no way to get rid of it,
unless we get rid of the entire classloader. This is effectively a
metaspace leak and people who create many solvers within the same JVM
run into an OOP relatively quickly.

The JVM guys are aware of this issue and have recently changed their
position on it. They argue it is not a bug, and that it only exists
because we are using an API we shouldn't have been using. According to
them, when used in the proper context, this issue doesn't happen. The
only relevant solution for us, however, is to remove the use of the API
altogether.

(For more details on the Java issue, see
https://bugs.openjdk.org/browse/JDK-8302154.)

The biggest performance drop as a result of this change was seen in
Cloud Balancing incremental score calculator. (-6 %.) This is not
surprising, because with its vastly superior speed, it is the most
exposed to these low-level calls. On the other hand, CS shows overall
positive results (+2 % median), with two outliers worsening by 4 %.

These impacts would likely be less visible in a real solver run as
opposed to a micro-benchmark focused on this. I think the trade-off is
worth it.