feat: Add concat operation to constraint streams
The concat operation acts as an or between two constraint streams,
propogating both stream tuples' downstream. It appears update/retract
may be called before an insert, meaning the out tuple stored inside
the in tuple is not set. As such, for update, if the out tuple is
not set, it does an insert instead, and for retract, it simply
ignores the operation.