fix: Update contents of concat's cloned tuple on update
Previously, concat did not update the content of the cloned tuples
in update. The `groupBy` and `map` building blocks both modify the contents
of their tuples instead of creating new ones. As a result, concat was
passing stale tuples with incorrect content downstream when placed
after a `groupBy` or `map` building block. Now, the contents of the
tuples are updated in update to prevent this issue.

Co-authored-by: Lukáš Petrovický <lukas@petrovicky.net>