fix: Deep clone unknown classes in GIZMO solution cloner (#341)
If a deep cloned class that is not accessible from the
SolutionDescriptor (for instance, a class that extends a @PlanningEntity
but adds no new variables) is inside the solution, it would incorrectly
be shallowly cloned. The GIZMO solution cloner will now fall back to the
reflection solution cloner if it encounters an unknown class.

---------

Co-authored-by: Lukáš Petrovický <lukas@petrovicky.net>
(cherry picked from commit 471a959a5dadbc2e02662464afa652f3a9c55ed5)