fix: Also include ConstraintCollector type in equality
Since all default constraint collectors of a given cardinality are
instances of the same Java class, we cannot rely on the instanceof
to make sure count() != countLong() (since they are instances of
the same class!). Added a new manatory arg ConstraintCollectorKind
to the default constraint collector constructor, so the collectors
can be distingished from one another.