fix: Support repeatable annotations like @ShadowVariable in Quarkus
Jandex, being Jandex, likes to fails sliently without telling you
anything. The culprit it this case is IndexView::getAnnotations.
IndexView::getAnnotations will appear to work, returning a collection
of AnnotationInstances. But it will sliently fail and return an
empty list when the annotation repeats on a target. The fix is to
use IndexView::getAnnotationsWithRepeatable instead.

Also did some cleanup in the generated Gizmo code, removing
meaningless try blocks, and not doing reflection to generate
the method descriptors.