test: do not create a new entity in EnvironmentMode test
When the test creates a new entity in its undo move and assigns it,
it increases the init score from 0 to 1, an impossible state that
should never occur in correctly implemented move(s). This could
result in a fail-fast in the future that is not related to the
EnvironmentMode being tested. Now the test use an existing but
different entity to achieve the desired corruption.