feat: Support multiple SolverManager instances (#564)
This pull request updates the Quarkus extension to support multiple
instances of Solver in an application. The new logic utilizes a map to
keep track of the configurations, and nothing changes for the current
behavior for single solvers, which are mapped by default to the key name
`default`.

A new build step that executes at runtime has been included to fetch an
updated properties instance from `TimefoldRuntimeConfig`. The reason we
need to register the managed beans at runtime is because
`Arc.container().instance()` does not return an instance of
`TimefoldRuntimeConfig` when using interfaces instead of classes.
Defining configuration properties as interfaces is the only way to use
`@WithUnnamedKey`. This is the default approach documented in both
Quarkus and Smallrye pages.