perf: Wait until calculateScore() to update changed entities in a Bavet session
Bavet cannot take advantage of variable infomation to only update some
nodes, so updating it on afterVariableChanged is pointless and can
degrade performance if an entity has multiple geninue or shadow
variables. Instead, we can record what has changed in
afterVariableChanged, and then update the entities in
calculateScore. This should improve performance for models
with multiple shadow/geninue variables.