fix: Pass all subclasses of Throwable to the exception handler in SolverManager
Previously, if a part of solving throws an Error (an unchecked exception that
normally should not be caught), the SolverManager will not pass
the error to the exception handler (nor will it log it to the console),
making the solver appear to be hanging.

Now, all subclasses of Throwable are passed to the exception handler.