chore: Use ObjectMapper instead of PojoInliner to transfer SolverConfig to generated method in spring
Finding out why ObjectMapper wasn't before was trickly. The short of it
is that java.lang.Class was registered for reflection, which made
GraalVM think its cachedHashCode field was reachable in an initializer,
which is not allowed. So we needed to make registerTypeRecursively
ignore java.lang.Class and java.lang.ClassLoader. Then ObjectMapper is
happy. I would use SolverConfigIO/JAXB, but JAXB accesses itself using
reflection, so I would need to register a bunch more (possible unstable)
classes for reflections to use it.