test: Create test for updating inverse relation variables on entities that supports pinning
Currently, doing an update on a value with a null inverse relation
variable but is assigned to an entity causes an IllegalStateException
to be thrown in a forEach node, even if only shadow variables are
updated. It requires the model to supports pinning, since it is
caused by a branch that is only triggered if the model supports
pinning.