chore: Change problem scale definition, fix flakey test
Problem scale is now defined as an approximation of the
base maximumValueSize log of the solution space.
This number is relatively close to the original calculation,
and has a particular meaning that can be used to estimate the
total search space. The log is rounded to 2 decimal places
and stored as a long.