chore: Add internal interface for planning cloning atypical types
Sometimes, the domain model has atypical types for fields (for instance,
UserList instead of List or ArrayList). The default planning cloner
originally failed fast when the field cannot be assigned a typical
value. Now, it checks if the collection implements the internal
interface. If it does, it uses it to create an empty collection/map.
This allows the default planning cloner to be used when the user
has control over UserList.