chore: remove many Swing-based examples (#730)
The entire chapter of use cases and examples was removed from the docs.
We do not want to advertise Swing-based examples.

Much of the NQueens and Cloud Balancing content was converted to School
Timetabling. A very small portion of content was removed without replacement.

NQueens and Cloud Balancing are still occasionally used in the docs to
explain some concepts. (They are simple enough to show well in
pictures.) Those explanations do not require any code, as they are
mostly concepts and visualizations. We may still decide to change those
pictures eventually.