fix: Benchmark duplicated metrics (#761)
This pull request resolves a bug in the benchmark module where metrics
are duplicated when running it multiple times, leading to inconsistent
results.

The problem is caused by a combination of an existing benchmark config
that is created after the first run and the property
`inheritedSolverBenchmark`. Therefore, the method
`ai.timefold.solver.benchmark.impl.DefaultPlannerBenchmarkFactory#buildEffectiveSolverBenchmarkConfigList`
will duplicate the values when inheriting the configuration at the line
`#163`.

The proposed solution fixes the problem by avoiding the duplicates in
the config class `ProblemBenchmarksConfig`.