fix: Do not throw an exception when the solution or entity classes are interfaces
The superclass of an interface is `null`. This caused this line of code
to throw an exception for solution classes that are interface:

```java
var superclass = bottomClass.getSuperclass();
lineageClassList.addAll(getAllAnnotatedLineageClasses(superclass, annotation));
```

Since `getAllAnnotatedLineageClasses` expected superclass to not be `null`.
`getAllAnnotatedLineageClasses` now returns an empty list for `null`
arguments.