fix: Do not fail on solution/entity classes as interfaces (#934)
The superclass of an interface is `null`. This caused this line of code
to throw an exception for solution classes that are interface:

```java
var superclass = bottomClass.getSuperclass();
lineageClassList.addAll(getAllAnnotatedLineageClasses(superclass, annotation));
```

Since `getAllAnnotatedLineageClasses` expected superclass to not be
`null`. `getAllAnnotatedLineageClasses` now returns an empty list for
`null` arguments.