feat: constraint weight override (#946)
Planning solution gets an optional field of type `Map` (or more likely a
custom map), with keys being the `constraintName` from
`asConstraint(constraintName)`, and the value being the constraint
weight.
Using the map is optional. Constraint weights not present in the map use
their default constraint weights as defined by `penalize(...)`.
Constraints present in the map but not present in the
`ConstraintProvider` trigger a fail-fast.

`asConstraintDescribed(constraintName, constraintDescription)` is
introduced to keep constraint documentation where it belongs, with the
constraints.

`ConstraintWeight`, `ConstraintConfiguration` and
`penalizeConfigurable(...)` are deprecated. They continue working, but
the new system doesn't use them.

---------

Co-authored-by: lee-carlon <lee@timefold.ai>
Co-authored-by: Frederico Gonçalves <zepfred@users.noreply.github.com>
Co-authored-by: Christopher Chianelli <christopher@timefold.ai>