fix: allow @PlanningId to be defined on interfaces (#1060)
Previously, `getAllMembers` excluded members defined in interfaces.
Presumably, this was because methods would have implementations in the
implementing class.
However, there is a flaw in that logic: implementations of a method do
not inherit the annotations of the declared method. As a result, when
LookupManager looks up the accessor of a class whose @PlanningId is
defined in an interface,
it finds none.

`getAllMembers` is changed to add all interface methods to the member
list, thus allowing interface members to be used in lookups.