fix: Pop off result of __setattr__ and __delattr__ in generic case
The cause of the VariableListener bug was discovered; when field
infomation is not available (due to cyclic references), the generic
__setattr__ and __delattr__ implementation forgot to pop off the
result of their method calls (recall: all Python functions have
a result). This cause an inconsistent stack size when
attributes without field infomation are assigned in a loop.

The fix is simple enough, just pop off the result.

This fix also reveals a bug in Python 3.10 for Enum that
need to be fixed.