docs: restructure introduction (#1151)
Brought the documentation more in line with the other documentation
elements.

- Shortened the introduction page
- Introduced the same PlanningAI Concepts page as on other sites (minus
the reference to the platform). This was originally also in the
"introduction"
- Introduced an FAQ to place information which people will look for
(like the licensing). Good place to add additional things Sales get
asked about often.

Also removed from the introduction page:

- The list of different Use Cases. While that is somewhat valuable, i'd
like to re-introduce those in the form of images. I left the current
image which already shows 4 use-cases. Longer term, let's introduce an
"Example Use Cases" heading to the documentation where we make these
shine a bit more.
- Note on Backwards Compatibility. I think this does not belong in an
Intro. I have removed it for now as I think the expectation is always
that this information should flow through the Changelog.
- How to use Timefold Solver => Should be part of the quickstarts, so
pushing people to the quickstarts sooner.