chore: unify the handling of list variable shadows (#1228)
All shadow variables related to a list variable are now handled in one
unified loop, as opposed to each having its own variable listener.
This should technically be a performance improvement, but in fact a
barely significant regression of 3 % was measured in some cases. (Other
cases show minor improvements.)

This change also makes internalized previous/next shadows possible, and
that in turn will help with the revamp of shadow variables that is
currently in development.
Therefore we introduce this change even though it may bring a small
performance regression.