fix: Recurse through @DeepPlanningCloned classes to discover all @DeepPlanningCloned classes
The GizmoSolutionClonerImplementor requires a set of all
@DeepPlanningCloned classes so it can create cloners for all of them.
GizmoCloningUtils.getDeepClonedClasses incorrectly assume all
@DeepPlanningCloned classes are directly accessible from either
the @PlanningSolution class or an @PlanningEntity class. However,
a @DeepPlanningCloned class might be referenced from another
@DeepPlanningCloned class, which would not be detected.

In order to detect all @DeepPlanningCloned classes, when
a class that is @DeepPlanningCloned is found and was not
processed yet, it is added to the to process queue.

Fixes #1208.