feat: allow phase-level terminations on solver-level (#1431)
Allows terminations which were previously only possible on phases (such
as diminished returns termination) to also be possible on solver level.
If a globally configured phase-level termination is not applicable to a
particular phase, it is skipped.
If a phase-level termination is configured on a phase to which it is not
applicable, the solver fails fast.

To achieve all of this, terminations are significantly refactored. There
is now a distinction between phase terminations, solver terminations,
and terminations that are both (universal).