fix: Make solver and phase terminations check if the solver was terminated early
Fixes a regression where terminateEarly was not checked until the end
of a phase. Previously, the top level termination was always BasicPlumbingTermination,
which checks for/sets the terminateEarly flag. However, with the
refactor to terminations, phases use different termination instances
(that are not instances of BasicPlumbingTermination). None of these
terminations check terminateEarly before, since that was the job of
the BasicPlubingTermination.

The abstract base classes of the terminations was changed:

- A new abstract method
  isPhaseTerminateConditionMet/isSolverTerminateConditionMet
- A final implementation of isPhaseTerminated/isSolverTerminated
  that checks if the solver is terminated, and then call the subclass
  method.

The reason our tests did not catch this is that termination is also
checked at the end/between phases, and that check works. The old
terminate early tests used a custom phase command that unblocks
a runner that terminates the solver. The issue is the solver is
terminated before LS (i.e. either between custom & CH, or during
CH), and thus the between check terminates the solver. To correctly
test for this, a custom easy score calculator that throws once
all entities are assigned is used for CH, and an event listener
that count LS steps is used for LS.