fix: phases should always check if the solver terminated early (#1456)
Fixes a regression where terminateEarly was not checked until the end of
a phase.
Previously, the top level termination was always
BasicPlumbingTermination, which checks for/sets the terminateEarly flag.
However, with the refactor to terminations, phases use different
termination instances (that are not instances of
BasicPlumbingTermination).
None of these terminations check terminateEarly before, since that was
the job of the BasicPlumbingTermination.

We fix that by making plumbing terminations always check for solver
termination.

Note: this PR is #1455 with additional tests.

---------

Co-authored-by: Lukáš Petrovický <lukas@petrovicky.net>