chore: cached dataset generation for move streams (#1349)
Demonstrates the integration of Bavet into cached Dataset generation.
It generifies Bavet and splits it into two implementations: a
constraint-oriented one, and a dataset-oriented one. The former remains
unchanged, the latter will be used in future move streams.

The new implementation is also feature-poor, only supporting uni streams
with filter and ifExists for now. The goal is not to expose the entire
Bavet functionality - instead, we intend to showcase that Bavet can be
integrated. As proof, see `UniDataStreamTest`.

Finally, we bring some early sketches of the entire Move Streams API.
The JIT skeleton, plus some of the generic moves. This code is currently
not functional, but it is next on the list on the road to executing a
first successful change move.