chore: make init score inaccessible to users (#1501)
The init score exposed by the Score interface is now always zero. To
find out when the solver has finished initializing a solution, users are
encouraged to catch the initialization event thrown by the solver
manager.

Internally, the init score remains. A new type was created (InnerScore),
which now carries the information that init score used to carry. All
code which requires to know the initialization status of the solution
now has to use InnerScore. Code which does not care about this
information can continue using Score.