fix: Update looped status of entity even if no variables changed
We previously only added a non-looped entity to affected entities
if one of its declarative variables changed. However, if null is
a valid value (such as for unassigned entities), then it might be
the case the looped status of an entity changed from looped to
non-looped. We now add an entity to the affected entities list if
its loop status changed.