fix: Benchmarker best score graph no longer empty (#1636)
During the init score removal (cf87aa7), we did not notice that this
impacts metrics as well. Benchmarker reads its data from metrics, and
therefore lost the information on what score is or is not initialized.
As a result, the default value was always false, resulting in empty
score graph.

This commit fixes that issue by registering a metric for the unassigned
count, which will be reused by benchmarker to provide the correct
initialization state.