fix: Graph now notify updater when cycles are formed or broken
Previously, the updater only considered nodes that were changed from
an edge insertion or deletion. However, an edge insertion/deletion
can also change the looped status of a disconnected node.
For example, if we have a -> b and b -> a, and we
remove a -> b, only b is marked as changed. This is incorrect, since
a changed indirectly, since it is no longer looped. Thus, the
graph implementations now mark any nodes whose SCC change from
singleton to multinode (looped) and vice-versa.