fix: process each entity only once and in topological order
Previously, the code relied on eventually consistency to avoid collections
(that is, sometimes suppliers will be called with stale data, but will
always eventually be called with correct data).
Now, an initial sort is done to ensure no data is stale, and a map
and comparator are utilized to prevent unnecessary computations.