perf: Use a single node to represent multiple shadow variables in graphs
When there is only a single declarative entity and all
non-declarative parents go in the same direction (ex: previous),
then some variables can share the same node in the graph. In
particular, there is a static topological order when only the
variables on a single entity is considered, and as long as
the next declarative shadow variable does not have any new
non-declarative parents (i.e. group), it can be grouped into
the same node as the previous one.