fix: centralize the ownership of ValueRangeManager to avoid subtle bugs (#1716)
`ValueRangeManager` is now owned and exclusively instantiated by
`ScoreDirector`.
Therefore all access to it is only possible within the context of a
working solution.
In order to achieve this, Move Streams get a special kind of filtering,
with access to `SolutionView` - this will be useful anyway, since I've
been meaning to do this at a later stage of Move Streams anyway.