feat: allowing to specify an alignment key to eliminate redundant calculations (#1718)
If a calculation only involves facts/variables on indirect variables,
then that calculation will result in the same value for entities that
share the same indirect variables/facts. To prevent redundant
recalculations, the user can specify an alignment key, and entities that
share the same group key for a shadow will only be calculated once.