test: remove race condition by making both branches return same value in SolverManagerTest
There was a race condition in SolverManagerTest#solveWithTerminationUnimprovedSpentLimit.
It calls UnimprovedTimeMillisSpentTermination#calculateTimeGradient, which returns
0.0 if the current phase does not send best solution events, or
solve duration since last best solution/limit if it does.
The test was relying on finishing before the Solver got to LocalSearchPhase, which is unreliable.
To fix it, we mock the clocks the Solver uses, so it always see solve duration since last
best solution as 0.