test: remove race condition in SolverManagerTest (#1831)
There was a race condition in
SolverManagerTest#solveWithTerminationUnimprovedSpentLimit. It calls
UnimprovedTimeMillisSpentTermination#calculateTimeGradient, which
returns 0.0 if the current phase does not send best solution events, or
solve duration since last best solution/limit if it does. The test was
relying on finishing before the Solver got to LocalSearchPhase, which is
unreliable.