fix: auto include entities reachable from the given entities
Let a, b be entities, and let a.start = b.end + 1.
Previously, if you were to do

    constraintVerifier.verifyThat(...).given(a)

or

    SolutionManager.updateShadowVariables(Solution.class, a)

an IllegalArgumentException would be thrown, since a references
a non-given entity b, which is unknown to the ShadowVariableGraph.

Now, we discover all entities referenced by the given entities and
automatically add them to the graph.

Note: This will also update the discovered
referenced entities shadow variables, which might be unexpected;
an alternative is to fail fast if we discover any new entities
to catch errors such creating multiple "copies" of the same entities
(i.e. value equal but not id equal).