fix: fail fast when referenced entities are not included in ConstraintVerifier.given or updateShadowVariables (#1836)
Let a, b be entities, and let a.start = b.end + 1. Previously, if you
were to do

    constraintVerifier.verifyThat(...).given(a)

or

    SolutionManager.updateShadowVariables(Solution.class, a)

an non-descriptive IllegalArgumentException would be thrown, since a
references a non-given entity b, which is unknown to the
ShadowVariableGraph.

Now, we discover all entities referenced by the given entities, and how
they are referenced, and include them in the exception message.