fix: process deep-cloned entities in a queue for GizmoSolutionCloner (#1849)
Previously, when a class is deep-cloned, a method that constructs and fully initializes a clone of that class is called. This has the potential to cause a StackOverflowException when a deep-cloned class refers to many other deep-cloned classes, such as in chained models.

Now, when a field is deeped cloned, it adds the field initializer to a cloneQueue, and that cloneQueue can only be processed for the first entity being cloned. As such, stack-height does not grow indefinitely.

Also changes method isInvalid to getIsInvalid so the GizmoSolutionCloner recongizes it as the getter of the corresponding field, and finally performs various cleanups in the GizmoSolutionCloner.