chore: refactor indexing to prepare for Neighborhoods (#1942)
Introduces a new collection that will serve as indexing backend for Neighborhoods. 
It is on average 10 - 15 % slower than the CS indexing backend, 
but gives us the power of direct access. This is a worthwhile trade-off, 
because Neighborhoods will need to access elements of the index randomly, 
and that would require copying the linked list to an array list, 
killing performance entirely.