fix: always update successors of single directional graph when non-declarative variables of parent are referenced
SingleDirectionalParentVariableReferenceGraph assumes that if no shadows of a changed entity is changed,
then no shadows of its successors can change (unless the successor had a variable that changed).
This is incorrect; consider [TestdataMixedValue("a"), TestdataMixedValue("b")];
if a.delay changes from 1 to 2, then its previousDelay is unchanged (as it has no previous),
but b.previousDelay should change from 1 to 2.

SingleDirectionalParentVariableReferenceGraph now always update
successors in the case that a declarative variable depends on a
non-declarative variable in its parent.