perf: improve memory use of entity value ranges (#1900)
This PR enables sorting the value ranges at the `ValueRangeManager`
level.

The sorting feature is likely useful during the CH phase, and the new
approach can address the sorting requirement for the existing CH
configurations. However, the existing selector sorting nodes remain
available, as they are required for range filtering or nearby selectors.

---------

Co-authored-by: Lukáš Petrovický <lukas@petrovicky.net>