chore: improve Neighborhoods design for scalability (#1903)
Previously, Neighborhoods depended on joins on the left-hand side -
these joins are, however, possibly massive. With large datasets, joining
100k entities with themselves would result in billions of entries per
move. This is clearly not acceptable and would never scale.

The new approach only uses joins within `pick()`, and these joins will
be enumerated on-demand. Therefore the memory use of this mechanism will
be minimized. (Although possibly still significant, especially for
random selection.)

The extra benefit of this approach is that it actually respects the
order of entities and facts that was given by the user.