perf: improve runtime performance of value ranges (#1959)
Makes two independent changes:

- Deduplication doesn't actually require creating a bitset. We can
compare the value ranges directly. With the hash trick that was already
there, it will be equally cheap and faster on account of dropping the
`BitSet`.
- The array list was very inefficient. It required creating a list of
`Integer` (not `int`, therefore boxing all the time), and even that list
was unnecessary as we can perform the same operation on the fly.